pyproject.toml
[project]
name = "fastapi-demo-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi[standard-no-fastapi-cloud-cli]>=0.124.4",
    "passlib>=1.7.4",
    "python-dotenv>=1.2.1",
    "sqlmodel>=0.0.27",
]
